Fix warnings in models

+++ models/src/main/scala/coop/rchain/models/rholang/sorter/ScoreTree.scala
@@ -84,7 +84,7 @@ trait ScoreTree {
   object Node {
     // Shortcut to write Node(1, Leaf(1)) instead of Node(Seq(Leaf(ScoreAtom(1)), Leaf(ScoreAtom(1))))
     def apply(left: Int, right: Tree[ScoreAtom]*): Tree[ScoreAtom] =
-      new Node(Seq(Leaf(left)) ++ right)
+      new Node(Seq(Leaf(left.toLong)) ++ right)
   }
 
   object ScoredTerm {
